+++ lib/advert/ui/pages/main_page/main_page.dart
@@ -8,7 +8,7 @@ import 'package:myecl/advert/providers/advert_provider.dart';
 import 'package:myecl/advert/providers/announcer_provider.dart';
 import 'package:myecl/advert/providers/is_advert_admin_provider.dart';
 import 'package:myecl/advert/ui/pages/advert.dart';
-import 'package:myecl/advert/ui/router.dart';
+import 'package:myecl/advert/router.dart';
 import 'package:myecl/advert/ui/tools/announcer_bar.dart';
 import 'package:myecl/advert/ui/tools/advert_card.dart';
 import 'package:myecl/tools/ui/refresher.dart';
